28.08.2022, update pdfexport skript
