--- pyTiBa.extension/pyTiBa.tab/Export.panel/exportPdf.pushbutton/dlgval.pkl
@@ -1,36 +1,0 @@
-(dp0
-Vprintername
-p1
-VPDFCreator
-p2
-sVprintername
-p3
-VPDFCreatorRevit
-p4
-sVpdfexport
-p5
-I1
-sVprintfilepath
-p6
-VC:\u005cUsers\u005ctill_laptop\u005cdesktop
-p7
-sVoutput
-p8
-I01
-sVoutput
-p9
-I0
-sVdwgexport
-p10
-I0
-sVpdfexport
-p11
-I01
-sVdwgexport
-p12
-I00
-sVprintfilepath
-p13
-VC:\u005cUsers\u005ctb_y530\u005cdesktop
-p14
-s.
