--- pyTiBa.extension/pyTiBa.tab/Export.panel/exportPdf.pushbutton/tb_SelectFromCheckboxes.xaml
@@ -1,123 +1,0 @@
-<Window xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-        mc:Ignorable="d"
-        ShowInTaskbar="True"
-        MinHeight="800" MinWidth="600"
-        HorizontalContentAlignment="Center" d:DesignWidth="498" d:DesignHeight="645.55">
-    <Window.Resources>
-        <Style x:Key="ClearButton" TargetType="{x:Type Button}">
-            <Setter Property="Background" Value="White"/>
-        </Style>
-    </Window.Resources>
-    <DockPanel Margin="10">
-        <DockPanel DockPanel.Dock="Top" Margin="0,0,0,10">
-            <TextBlock FontSize="14" Margin="0,3,10,0"
-                        DockPanel.Dock="Left"><Run Text="Filter:"/></TextBlock>
-            <StackPanel>
-                <TextBox x:Name="search_tb" Height="25px"
-                            TextChanged="search_txt_changed"/>
-                <Button Style="{StaticResource ClearButton}"
-                            HorizontalAlignment="Right"
-                            x:Name="clrsearch_b" Content="x"
-                            Margin="0,-25,5,0" Padding="0,-4,0,0"
-                            Click="clear_search"
-                            Width="15px" Height="15px"/>
-            </StackPanel>
-        </DockPanel>
-        <StackPanel DockPanel.Dock="Bottom">
-            <Grid>
-                <Grid.RowDefinitions>
-                    <RowDefinition Height="Auto" />
-                    <RowDefinition Height="Auto" />
-                </Grid.RowDefinitions>
-                <Grid.ColumnDefinitions>
-                    <ColumnDefinition Width="226*" />
-                    <ColumnDefinition Width="267*" />
-                    <ColumnDefinition Width="247*" />
-                </Grid.ColumnDefinitions>
-                <Button x:Name="checkall_b"
-                            Grid.Column="0" Grid.Row="0"
-                            Content="Check" Click="check_all"
-                            Margin="0,10,3,0"/>
-                <Button x:Name="uncheckall_b"
-                            Grid.Column="1" Grid.Row="0"
-                            Content="Uncheck" Click="uncheck_all"
-                            Margin="3,10,3,0"/>
-                <Button x:Name="toggleall_b"
-                            Grid.Column="2" Grid.Row="0"
-                            Content="Toggle" Click="toggle_all"
-                            Margin="3,10,0,0"/>
-            </Grid>
-            <Button x:Name="select_b" Content="Print Selection"
-                            Click="button_select" Margin="0,10,0,0"/>
-            
-                <GroupBox Margin="0,5,0,0">
-                    <StackPanel>
-                        <TextBox x:Name="txtbox_paranames" Margin="5,10,10,10" Height="20px" />
-                        <TextBlock Margin="2,0,0,2">
-                              <Run Text="INPUT: ParameterNames/SeparatorChars (_, ., -, blank) separated by comma "/><LineBreak/>
-                              Example: Sheet Number,_,Current Revision,_,Sheet Name,_,date,_,time <LineBreak/>
-                              Chars not allowed: \ / : &quot; ? * | &gt; 
-                        </TextBlock>
-                        <TextBlock Margin="2,0,0,5">
-                              <Bold>keywords:</Bold> date -> 25-08-19, time -> 10.34
-                              <LineBreak/> %d: day,  %m: month,  %y: year,  %H: hour,  %M: minute
-                        </TextBlock>
-                        <StackPanel Orientation="Horizontal">
-                            <Button x:Name="preview_b" Content="Preview" 
-                                Click="preview_click" Margin="5,5,5,0" Width="50" Height="20" 
-                                HorizontalAlignment="Left"/>
-                            <Label x:Name="label_preview" Margin="20,5,5,0" > 
-                                <!-- Textbox as Content of Label to show underscore letter: -->
-                                <TextBox x:Name="lb_txtbox_preview" Text="" BorderThickness="1" />
-                            </Label>
-                        </StackPanel>
-                        <StackPanel Orientation="Horizontal">
-                            <Label Content="PrinterName:" /> 
-                            <TextBox x:Name="txtbox_printername" Margin="5,5,5,5"
-                                     Width="150" Height="20" />
-                        </StackPanel>
-                        <StackPanel Orientation="Horizontal">
-                            <Label  Content="PrintPath:       "/>
-                            <Button x:Name="but_printfilepath" Content="Select"
-                                    Width="60" Height="20" Click="selectprintfilepath_click"/>
-                            <Label x:Name="lb_printfilepath"
-                                   Content=""/>
-                        </StackPanel>
-                    </StackPanel>
-                </GroupBox>
-            
-            <StackPanel Orientation="Horizontal">
-                <CheckBox x:Name="chbox_pdfexport"
-                            Margin="10,5,0,5"
-                            Content="Export as pdf"
-                            IsChecked="True" />
-                <CheckBox x:Name="chbox_dwgexport"
-                            Margin="10,5,0,5"
-                            Content="Export as dwg" 
-                            IsChecked="False" />
-                <CheckBox x:Name="chbox_output" 
-                            Margin="10,5,0,5" 
-                            Content="Show Output Window/Report" 
-                            IsChecked= "True" />
-                <Button HorizontalAlignment="Right" x:Name="b_savesettings" Content="Save Settings"
-                                    Margin= "20,5,0,5" Width="80" Height="20" Click="savesettings_click"/>
-            </StackPanel>
-        </StackPanel>
-        <ListView x:Name="list_lb">
-            <ListView.ItemTemplate>
-                <DataTemplate>
-                    <StackPanel Orientation="Horizontal">
-                        <CheckBox IsChecked="{Binding state}"
-                                Checked="check_selected"
-                                Unchecked="uncheck_selected" />
-                        <TextBlock Text="{Binding name}"
-                                    Margin="10,0,0,0"/>
-                    </StackPanel>
-                </DataTemplate>
-            </ListView.ItemTemplate>
-        </ListView>
-    </DockPanel>
-</Window>
