+++ pyTiBa.extension/pyTiBa.tab/Export.panel/exportPdf_v21.pushbutton/dlgval.pkl
@@ -1,0 +1,19 @@
+(dp0
+Vprintername
+p1
+VPDFCreatorRevit
+p2
+sVoutput
+p3
+I01
+sVpdfexport
+p4
+I01
+sVdwgexport
+p5
+I00
+sVprintfilepath
+p6
+VC:\u005cUsers\u005ctb_y530\u005cdesktop
+p7
+s.
